moving all synch_GHS related code to single file
